Create visitors & nodes subdirectories (#2313)
- 📦 Straigten up directories
- 🍀 Remove empty Leaf.cpp
