+++ src/solver/expressions/include/antares/solver/expressions/visitors/PrintVisitor.h
@@ -19,8 +19,25 @@
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
 #pragma once
-#include <antares/solver/expressions/AddNode.h>
-#include <antares/solver/expressions/ComponentNode.h>
-#include <antares/solver/expressions/Leaf.h>
-#include <antares/solver/expressions/NegationNode.h>
-#include <antares/solver/expressions/PortFieldNode.h>
+
+#include <antares/solver/expressions/visitors/Visitor.h>
+
+namespace Antares::Solver::Expressions
+{
+class PrintVisitor: public Visitor<std::string>
+{
+public:
+    using Base = Visitor<std::string>;
+
+private:
+    std::string visit(const AddNode& add) override;
+
+    std::string visit(const NegationNode& neg) override;
+
+    std::string visit(const ParameterNode& param) override;
+    std::string visit(const LiteralNode& lit) override;
+    std::string visit(const PortFieldNode& port_field_node) override;
+    std::string visit(const ComponentVariableNode& component_variable_node);
+    std::string visit(const ComponentParameterNode& component_parameter_node);
+};
+} // namespace Antares::Solver::Expressions
