+++ src/solver/expressions/nodes/PortFieldNode.cpp
@@ -18,21 +18,13 @@
 ** You should have received a copy of the Mozilla Public Licence 2.0
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
-#pragma once
-
-#include <memory>
-
-#include <antares/solver/expressions/Leaf.h>
+#include <antares/solver/expressions/nodes/PortFieldNode.h>
 
 namespace Antares::Solver::Expressions
 {
-class PortFieldNode: public Node
+PortFieldNode::PortFieldNode(const std::string& port_name, const std::string& field_name):
+    port_name_(port_name),
+    field_name_(field_name)
 {
-public:
-    explicit PortFieldNode(const std::string& port_name, const std::string& field_name);
-    virtual ~PortFieldNode() = default;
-    //   std::string
-    std::string port_name_;
-    std::string field_name_;
-};
+}
 } // namespace Antares::Solver::Expressions
