+++ src/solver/expressions/visitors/CloneVisitor.cpp
@@ -18,8 +18,8 @@
 ** You should have received a copy of the Mozilla Public Licence 2.0
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
-#include <antares/solver/expressions/CloneVisitor.h>
-#include <antares/solver/expressions/ExpressionsNodes.h>
+#include <antares/solver/expressions/nodes/ExpressionsNodes.h>
+#include <antares/solver/expressions/visitors/CloneVisitor.h>
 
 namespace Antares::Solver::Expressions
 {
